Move Bet class to a separate file
